Database migration + new controller
