+++ backend/Controllers/AuthController.cs
@@ -88,80 +88,4 @@ public class AuthController : ControllerBase
             Token = jwtToken
         });
     }
-
-    [Authorize]
-    [HttpPost("{userId}/change-username")]
-    public async Task<ActionResult> ChangeUsername(int userId, [FromBody] ChangeUsernameRequest request)
-    {
-        var tokenUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "0");
-        if (tokenUserId != userId)
-        {
-            return Forbid();
-        }
-        var success = await _authService.ChangeUsernameAsync(userId, request.NewUsername);
-        if (!success)
-            return BadRequest(ApiResponse.Error("Username already taken", ErrorCodes.USERNAME_EXISTS));
-
-        return Ok(ApiResponse.Ok("Username updated successfully"));
-    }
-
-    [Authorize]
-    [HttpPost("{userId}/change-password")]
-    public async Task<IActionResult> ChangePassword(int userId, [FromBody] ChangePasswordRequest request)
-    {
-        var tokenUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "0");
-        if (tokenUserId != userId)
-        {
-            return Forbid();
-        }
-        var success = await _authService.ChangePasswordAsync(userId, request.CurrentPassword, request.NewPassword);
-        if (!success)
-            return BadRequest(ApiResponse.Error("Invalid current password", "INVALID_PASSWORD"));
-        return Ok(ApiResponse.Ok("Password changed successfully"));
-    }
-
-    [Authorize]
-    [HttpPost("{userId}/request-email-change")]
-    public async Task<ActionResult> RequestEmailChange(int userId, [FromBody] ChangeEmailRequest request)
-    {
-        var tokenUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "0");
-        if (tokenUserId != userId)
-            return Forbid();
-
-        var success = await _authService.SendEmailVerificationAsync(userId, request.NewEmail);
-        if (!success)
-            return BadRequest(ApiResponse.Error("Email already taken or invalid", ErrorCodes.EMAIL_EXISTS));
-
-        return Ok(ApiResponse.Ok("Verification email sent. Please confirm your new email address."));
-    }
-
-    [HttpPost("confirm-email-change")]
-    public async Task<ActionResult> ConfirmEmailChange([FromBody] TokenRequest token)
-    {
-        var success = await _authService.ConfirmEmailChangeAsync(token.Token);
-        if (!success)
-            return BadRequest(ApiResponse.Error("Invalid or expired token", "INVALID_TOKEN"));
-
-        return Ok(ApiResponse.Ok("Email successfully changed and verified."));
-    }
-
-    [Authorize]
-    [HttpPost("{userId}/upgrade-plan")]
-    public async Task<ActionResult> UpgradePlan([FromRoute] int userId, [FromBody] UpgradePlanRequest upgradePlanRequest)
-    {
-        if (upgradePlanRequest.NewPlan == null || !Enum.IsDefined(typeof(SubscriptionPlan), upgradePlanRequest.NewPlan))
-        {
-            _logger.LogWarning("Requested upgrade plan value: {upgradePlanRequest}", upgradePlanRequest);
-            return BadRequest(ApiResponse.Error("Invalid subscription plan value"));
-        }
-
-        var tokenUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "0");
-        if (tokenUserId != userId)
-            return Forbid();
-
-        var success = await _authService.UpgradePlanAsync(userId, upgradePlanRequest.NewPlan.Value);
-        if (!success)
-            return BadRequest(ApiResponse.Error("Error upgrading plan"));
-        return Ok(ApiResponse.Ok("Plan upgraded successfully"));
-    }
 }
