+++ backend/Services/Interfaces/IAuthService.cs
@@ -49,16 +49,6 @@ public interface IAuthService
 
     Task<string?> ConfirmEmailAndGenerateTokenAsync(string token);
 
-
-    Task<bool> ChangeUsernameAsync(int userId, string newUsername);
-
-    Task<bool> ChangePasswordAsync(int userId, string oldPassword, string newPassword);
-
-    Task<bool> SendEmailVerificationAsync(int userId, string newEmail);
-
-    Task<bool> ConfirmEmailChangeAsync(string token);
-
-    Task<bool> UpgradePlanAsync(int userId, SubscriptionPlan subscriptionPlan);
     #endregion
 
 }
